1. refactor photo and user services; 2. add photo deleting function in user controller;
