+++ API/Controllers/UserController.cs
@@ -1,5 +1,7 @@
 using System;
 using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
+using System.Linq;
 using System.Security.Claims;
 using System.Threading.Tasks;
 using AutoMapper;
@@ -65,7 +67,7 @@ namespace API.Controllers
 
             if (result.Error != null) return BadRequest(result.Error.Message);
 
-            var photo = await _photoService.AddPhotoAsync(result, user);
+            var photo = await _userService.AddPhotoAsync(result, user);
 
             if (photo != null) 
             {
@@ -81,9 +83,31 @@ namespace API.Controllers
         [HttpPut("set-main-photo/{photoId}")]
         public async Task<ActionResult> SetMainPhoto(int photoId)
         {
-            var result = await _photoService.SetMainPhotoAsync(User.GetUsername(), photoId);
+            var result = await _userService.SetMainPhotoAsync(User.GetUsername(), photoId);
 
             return result ? NoContent() : BadRequest("Fail to set main photo");
         }
+
+        [HttpDelete("delete-photo/{photoId}")]
+        public async Task<ActionResult> DeletePhoto(int photoId)
+        {
+            var user = await _userService.GetUserByUsername(User.GetUsername());
+
+            var photo = user.Photos.FirstOrDefault(x => x.Id == photoId);
+
+            if (photo == null) return NotFound();
+
+            if (photo.IsMain) return BadRequest("You cannot delete your main photo");
+
+            if (photo.PublicId != null)
+            {
+                var result = await _photoService.DeletePhotoAsync(photo.PublicId);
+                if (result.Error != null) return BadRequest(result.Error.Message);
+            }
+
+            if (await _userService.DeletePhotoAsync(user, photo)) return Ok();
+
+            return BadRequest("Fail to delete the photo");
+        } 
     }
 }
