+++ API/Services/Interface/IPhotoService.cs
@@ -8,8 +8,6 @@ namespace Interfaces
     public interface IPhotoService
     {
         Task<ImageUploadResult> UploadPhotoAsync(IFormFile file);
-        Task<Photo> AddPhotoAsync(ImageUploadResult result, AppUser user);
-        Task<bool> SetMainPhotoAsync(string username, int photoId);
         Task<DeletionResult> DeletePhotoAsync(string publicId);
     }
 }
