+++ API/Services/Interface/IUserService.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using CloudinaryDotNet.Actions;
 using DTOs;
 using Entities;
 
@@ -11,5 +12,8 @@ namespace Services.Interface
         Task<MemberDTO> GetUser(string username);
         Task<AppUser> GetUserByUsername(string username);
         Task<bool> UpdateUser(MemberUpdateDTO memberUpdateDTO, string username);
+        Task<Photo> AddPhotoAsync(ImageUploadResult result, AppUser user);
+        Task<bool> SetMainPhotoAsync(string username, int photoId);
+        Task<bool> DeletePhotoAsync(AppUser user, Photo photo);
     }
 }
