+++ API/Services/PhotoService.cs
@@ -1,12 +1,9 @@
-using System.Linq;
 using System.Threading.Tasks;
 using CloudinaryDotNet;
 using CloudinaryDotNet.Actions;
-using Entities;
 using Helpers;
 using Interfaces;
 using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Options;
 
 namespace Services
@@ -29,23 +26,6 @@ namespace Services
 		_userRepository = userRepository;
     }
 
-    public async Task<Photo> AddPhotoAsync(ImageUploadResult result, AppUser user)
-    {
-		var photo = new Photo
-		{
-			Url = result.SecureUrl.AbsoluteUri,
-			PublicId = result.PublicId
-		};
-
-		if(user.Photos.Count == 0)
-		{
-			photo.IsMain = true;
-		}
-
-		user.Photos.Add(photo);
-
-		return await _userRepository.SaveAllAsync() ? photo : null;
-    }
     public async Task<ImageUploadResult> UploadPhotoAsync(IFormFile file)
     {
       var uploadResult = new ImageUploadResult();
@@ -62,23 +42,6 @@ namespace Services
       }
       
       return uploadResult;
-    }
-
-    public async Task<bool> SetMainPhotoAsync(string username, int photoId)
-    {
-      var user = await _userRepository.GetUserByUsernameAsync(username);
-
-      var photo = user.Photos.FirstOrDefault(x => x.Id == photoId);
-
-      if (photo.IsMain) return true;
-
-      var currentMain = user.Photos.FirstOrDefault(x => x.IsMain);
-      if (currentMain != null) currentMain.IsMain = false;
-      photo.IsMain = true;
-
-      if (await _userRepository.SaveAllAsync()) return true;
-
-      return false;
     }
 
     public async Task<DeletionResult> DeletePhotoAsync(string publicId)
