+++ API/Services/UserService.cs
@@ -1,6 +1,8 @@
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 using AutoMapper;
+using CloudinaryDotNet.Actions;
 using DTOs;
 using Entities;
 using Interfaces;
@@ -42,6 +44,48 @@ namespace Services
             _userRepository.Update(user);
 
             return await _userRepository.SaveAllAsync();
+		}
+
+		public async Task<Photo> AddPhotoAsync(ImageUploadResult result, AppUser user)
+		{
+			var photo = new Photo
+			{
+				Url = result.SecureUrl.AbsoluteUri,
+				PublicId = result.PublicId
+			};
+
+			if(user.Photos.Count == 0)
+			{
+				photo.IsMain = true;
+			}
+
+			user.Photos.Add(photo);
+
+			return await _userRepository.SaveAllAsync() ? photo : null;
+		}
+
+		public async Task<bool> SetMainPhotoAsync(string username, int photoId)
+		{
+			var user = await _userRepository.GetUserByUsernameAsync(username);
+
+			var photo = user.Photos.FirstOrDefault(x => x.Id == photoId);
+
+			if (photo.IsMain) return true;
+
+			var currentMain = user.Photos.FirstOrDefault(x => x.IsMain);
+			if (currentMain != null) currentMain.IsMain = false;
+			photo.IsMain = true;
+
+			if (await _userRepository.SaveAllAsync()) return true;
+
+			return false;
+		}
+
+		public async Task<bool> DeletePhotoAsync(AppUser user, Photo photo)
+		{
+			user.Photos.Remove(photo);
+
+			return await _userRepository.SaveAllAsync();
 		}
 	}
 }
